combine create/edit of networks
