+++ ui/src/utils/stringUtils.js
@@ -2,4 +2,4 @@ import { replace } from 'ramda';
 
 
 export const capitalize = str => str.charAt(0).toUpperCase() + str.slice(1);
-export const removeUnderscores = str => replace(/_/g, ' ', str);
+export const removeUnderscores = str => replace(/_/g, ' ', str); // TODO: not working
